refactor: move TRAVEL_PRODUCT_LIST to dummy file
